requirements.txt
tgcrypto
ffmpeg-python
python-dotenv
apscheduler
pytz
requests
telethon
aiofiles
aiohttp
pymongo
motor